Change Item key. Reverse previous changes. List re-rendering is fixed. UUID was creating a new key with every itemList state change, replaced with index for now.
