--- src/components/InputButton.jsx
@@ -1,8 +1,0 @@
-<button
-  onClick={() => {
-    props.onAdd(inputText);
-    setInputText("");
-  }}
->
-  +
-</button>;
